v2: net: ethernet: prueth: fix packet loss by using threaded irq handler
Currently using NAPI polling, overflow happens at the ingress queues
when iperf udp traffic is offered to the ingress port at 92Mbps with
MTU sized frames. As the ingress queues have limited buffers (only
6 MTU sized buffers), overflow happens as the NAPI polling is found
to be too slow. So fix this by processing the received frames as quickly
as possible using a threaded irq handler function.

This is an important requirement for supporting industrial Ethernet
such as the one used in substation automation as it requires a loss
free transfer of frames across the network.

Changes in v2:
- added one more queue check cycle

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Aaron Kramer <a-kramer@ti.com>
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>

%% original patch: 0027-v2-net-ethernet-prueth-fix-packet-loss-by-using-thre.patch