HACK: net: hsr: Optimize Tx path
The HSR code takes a single untagged frame, creates two copies of those, tags
them and passes them down to the both slave ports. This is all correct and works
as expected.

However, due to this software architecture, three skbs will be allocated and
free'd on the way. This results in a lot of cpu cycles. This can be optimized by
using only a single skb. The HSR code adds the HSR/PRP tags to the original skb
and modifies it as needed for the slaves. This single skb is then passed to both
slaves after each other and free'd by the HSR code. As a consequence only one
skb is used.

In order to make that work, there are a few assumptions:

 * The PRU driver doesn't access the skb data after .ndo_start_xmit()
 * The skb passed to HSR has enough head and tailroom set for the HSR tags

These assumptions are fulfilled, because the surrounding code has been modified
as well.

After applying this optimization, the time spent in sendto() systemcall is
reduced by 20us using six parallel streams.

Signed-off-by: Kurt Kanzenbach <kurt@linutronix.de>
Reviewed-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>