net: ethernet: ti: prueth: fix rx timestamp reading when packet wrapped buf boundary
The is one case when RX TS is not read properly:
 P - packet data
 T - timestamp
 ....PPPT - fail
 T....PPP - ok
 PT....pp - ok

In the above failed case, the packet will be detected as wrapped packet
while in fact it fully fits at the end of the queue buffer. The packet data
will be red properly and remaining=0, which will cause src pointer switch
to the beginning of the queue buffer and, so, random data will be red as
TS.

The similar will happen with packet without TS, but without packet data
corruption - just additional run-time overhead.
 ....PPPP

To fix such cases do not treat packets as wrapped packets if they fully
fits at the end of the queue buffer.

Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>

%% original patch: 0035-net-ethernet-ti-prueth-fix-rx-timestamp-reading-when.patch