blk-mq: Don't push complete operation to softirq
For single-queue hardware the blk-completion is deferred to softirq-context. If
the request completes already in process context then its processing will be
deferred to the ksoftirqd task. A runnable ksoftirqd will collect all other
pending/raised softirqs even if raised in a (threaded-)IRQ context.

This interacts badly with network related workload especially if one tasks
involved has a higher priority. The priority will lead to PI-boost via local-BH
lock. Due this PI-boost long running I/O will be served at higher priority.

To avoid this ping-pong the request can be completed directly in caller's
context. The softirq-context is used to limit the IRQ-off time if invoked from a
IRQ-handler.

Signed-off-by: Kurt Kanzenbach <kurt@linutronix.de>
Reviewed-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>

%% original patch: 0001-blk-mq-Don-t-push-complete-operation-to-softirq.patch