net: ethernet: ti: ale: keep track of max index in table entries
The size of the CPSW ALE table entries is multiple of 1024. There are
several functions in the driver that loop those 1024 entries in order to
add, search, modify, delete entries in the table. Hence these loops in
the driver could take an unnecessary amount of CPU time in most cases
where the number of table entries is low.

Add a new variable in cpsw_ale_params to keep track of the highest index
added in the table entries and change all the loops in the driver to
this index instead of the max entries.

Note: only the loop to search for a free entry should be able to go
through the max entries.

Signed-off-by: Anthony Harivel <anthony.harivel@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>