Merged PR 3423: Introduce device tree file for RTU530 530CPM02 CMU board in PG Yocto Linux code base
With this pull request the device tree file for RTU530 530CPM02 CMU board is added to the PG Yocto Linux code base. The 530CPM02 board suitable for multi-CMU setups was missing in the code base.

Besides the added file, the copyright header of the existing RTU530 device tree files are corrected according to the RTU500 firmware definition. Furthermore, the obsolete company name ABB was removed.

Signed-off-by: Piotr Lewicki <piotr.l.lewicki@hitachienergy.com>

%% original patch: 0006-Merged-PR-3423-Introduce-device-tree-file-for-RTU530.patch