Merged PR 5358: posix-timers: Preserve return value in clock_adjtime32()
NOTE: This is not a complete solution but "good enough", further upstream solution have correct return code for both get/put_old_timex32().

posix-timers: Preserve return value in clock_adjtime32()

commit 2d036dfa5f10df9782f5278fc591d79d283c1fad upstream.

The return value on success (>= 0) is overwritten by the return value of
put_old_timex32(). That works correct in the fault case, but is wrong for
the success case where put_old_timex32() returns 0.

Just check the return value of put_old_timex32() and return -EFAULT in case
it is not zero.

[ tglx: Massage changelog ]

Fixes: 3a4d44b61625 ("ntp: Move adjtimex related compat syscalls to native counterparts")
Signed-off-by: Chen Jun <chenjun102@huawei.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Reviewed-by: Richard Cochran <richardcochran@gmail.com>
Cc: stable@vger.kernel.org
Link: https://lore.kernel.org/r/20210414030449.90692-1-chenjun102@huawei.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Piotr Lewicki <piotr.l.lewicki@hitachienergy.com>

%% original patch: 0014-Merged-PR-5358-posix-timers-Preserve-return-value-in.patch