With the Linux real-time patch (RT patch) the low-level interrupt processing is performed in dedicated real-time threads per interrupt. This is true for the serial interfaces as well and each UART has a corresponding RT thread for the interrupt processing. The application functions to send and receive via the serial interfaces, are executed in RT threads as well. But between the interrupt processing and the application, is the Linux TTY layer using kernel workqueues. These workqueues are used for the TTY flip buffer handling and are processed by kernel threads from a thread pool. And all the threads from the kernel pool are running as nonreal-time standard threads.
As result, the receive path of the serial interfaces loses the
real-time priority and any other RT thread running can postpone the
time critical serial communication. This leads to receive timeouts
and connected I/O boards may become invalid depending on the CPU
utilization.

To solve this issue this commit introduces a device tree option to
enable dedicated kernel workqueue threads for the TTY buffer
handling. When enabled, the TTY port of the UART uses a dedicated
workqueue for flip buffer processing. The kernel workqueue thread
has RT priority and ensures real-time priority on the receive path.
The RT priority is set, as default, to the same value as for the
UART IRQ thread.

Besides the described TTY driver changes, the DMA (Direct Memory
Access) is disabled for all UARTs on the RTU530 CMU modules. This is
required because DMA doesn’t work reliable for the AM437x UARTs and
may result in unexpected kernel crashes.

Signed-off-by: Piotr Lewicki <piotr.l.lewicki@hitachienergy.com>

%% original patch: 0023-With-the-Linux-real-time-patch-RT-patch-the-low-leve.patch