net: dsa: mv88e6xxx: Improve indirect register access perf on 6393
When operating in multi-chip mode, the 6393 family maps a subset of
commonly used global registers to the outermost address space (in
which only the SMI command and data registers were available in
previous families).

Therefore, add a new set of SMI operations which remaps accesses to
such registers to the corresponding directly addressable register. All
other accesses use the regular indirect interface.