Merged PR 9449: Panic Handler Module
Module used to save kernel logs to persistent memory upon crash. This is triggered, and the logs are retrieved, through the kmsg_dumper callchain.
To use the module the parameter kmsg_dumper_memory has to be set at initialization, this specifies the area of memory to write to as well as the size of the area.
`insmod panic_handler.ko kmsg_dumper_memory=<destination-memory>,<size>`
Where `<destination-memory>` is the physical address specifying the start of the requested area, and `<size>` is the total allowed size of the buffer including header information (`sram_log_area`).
Related work items: #59539 #59540
https://dev.azure.com/PG-CT/PGLinux/_git/pg-linux/pullrequest/9449