net: bridge: drop delay for applying strict multicast filtering
This *local* patch drops the initial delay before applying strict multicast
filtering, introduced in [1] and recently updated in [2].

The main reason for this patch is RFC conformance and customer expectations.
At power on we assume the querier role by default and are expected to stop
unknown flooding as soon as we have a membership report in place, not after
10 seconds.

A proper fix for upstreaming could be to add a knob to disable the delay.

[1]: https://lore.kernel.org/netdev/1375311980-25575-1-git-send-email-linus.luessing@web.de/
[2]: https://lore.kernel.org/netdev/20240127175033.9640-1-linus.luessing@c0d3.blue/

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>