Merged PR 11008: [Fault Dumper] Add a lock to the dump process and set config to Bool
When multiple dumps are occuring at the same time there is a possibility of a race-condition which will corrupt the entry. Add a mutex to mitigate this, since this is a critical context we only try to get the lock and not wait until it is available.

Snprintf information into the string and not only the name, making the dump contain better information.

writel on the total dump size which validates the entire memory area, force a wmb so it writes it in order.

Related work items: #59539

Signed-off-by: Piotr Lewicki <piotr.l.lewicki@hitachienergy.com>

%% original patch: 0032-Merged-PR-11008-Fault-Dumper-Add-a-lock-to-the-dump-.patch