net: dsa: tag_dsa: Use tag priority as initial skb->priority
Use the 3-bit priority field from the DSA tag as the initial packet
priority on ingress to the CPU.

These bits contain the frame priority (FPri) of the packet. Barring
any non-standard policy configuration, this corresponds to:

1. The original PCP bits if the packet is received with a VLAN tag
2. ...or the the 3 MSBs of the DSCP field, if it is an IP packet
3. ...otherwise 0

Ideally we could then map this priority to an internal one, like we
can do with an "ingress-qos-map" on VLAN interfaces. Until that is
implemented, support the setup that is likely to be the most common; a
1:1 mapping from FPri to skb->priority.