net: dsa: mv88e6xxx: Add mqprio qdisc support
Add support for attaching mqprio qdisc's to mv88e6xxx ports and use
the packet's traffic class as the outgoing priority when no PCP bits
are available.

We cannot directly target a particular transmit queue on mv88e6xxx
chips, we can only specify the frame priority (FPri), which will
typically cause the switch to set the packet's queue priority (QPri)
to the same value.

For symmetry, we give precedence to an explicit PCP value available in
the packet, since that maps directly to the chip's FPri.

In cases where no PCP bits are available, use the packet's queue
mapping to derive its traffic class and use that as the FPri. Since
the packet will always be egressing without a tag in these cases, we
are effectively only specifying the QPri.

Since FPri is always a 3-bit field, even on older chips with only 4
physical queues, always report 8 queues and let the chip's policy
handle the mapping down to the "real" number.