net: dsa: mv88e6xxx: Add workaround for watchdog issue on 6393X
Before this change, in some edge condition, the queue controller would
inadvertantly trip the watchdog, without any event being
reported. This trapped the CPU in an endless loop, continuously
serving trying to serve the watchdog interrupt without ever managing
to clear the condition.

Work around the issue by increasing the size of the free pool FIFO,
presumably introducing a small time delay before a buffer can be
reused again.

This issue is not documented in any erratum or in the chip's release
notes, at least as of October 2024.