net: dsa: mv88e6xxx: Rate limit watchdog event logging
Avoid the kernel log filling up with redundant messages in the case
when the interrrupt is not serviced properly.

Elevate log severity to error, as the datapath is most likely about to
be reset. This will typically cause packets to be dropped, making _err
more appropriate than _info.

To further limit the amount of space in the log, collect all the
information in a single line.