net: dsa: mv88e6xxx: collapse disabled state into blocking
This patch changes the behavior of switchcore ports wrt. the port state.
Instead of disabling the port, the driver now treats the disabled state
the same as the blocking and listening state.

On this particular type of device the disabled state means no traffic is
passed in or out.  This interferes with protocols that perform any type
of link qualification.  E.g., an LACP link aggregate (bond) used as a
bridge port with member ports from an mv88e6xxx switch core.  Without
this patch, the bond is set to 'state disabled'. which is offloaded to
each member port, preventing LACPDUs from passing through to qualify the
link and become active.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>