Don't allow too low timestamp
This fix should be propagated to all of the branches.

This adds another limitation (the other one is timestamp >= uptime) to time-of-day timestamp to avoid a false positive, failing check in fsck when booting. This happens if a file is deleted when the time of day, in unix 'ticks', is lower than the maximum number of inodes in the filesystem.

To reproduce: date -s @20000 && touch testfile && rm testfile (and then pull the plug)

The reason for the check against the maximum number of inodes, is an overload of the deletion date attribute (i_dtime attribute). With orphaned inodes, the deletion date attribute is used for a linked orphaned list, overwriting the deletion date. So, the check against the maximum number of inodes is used, most probably, to determine whether the attribute is an actual deletion date or a link in the orphaned list.

The problem is fixed in the inode implementation, by checking whether the deletion date is below the maximum number of inodes and if yes setting the deletion date attribute to the maximum number + 1.

This PR is for now only to notify the interested parties that such fix exist.

This is related to https://dev.azure.com/PG-BLP/BIZ/_workitems/edit/410220