Merged PR 17006: Wave 3 release 3.8.0
Wave 3 release 3.8.0
The following Enhancement/fix has been included in this release.

Linux:
Fix for the CPU load increase when Linux PTP is running without affecting the tap points

Earlier, the tap API was commented for saving ~5 to 10% of CPU load when Linux PTP is running, and found that this is due to the socket type being used in Linux PTP.

After changing the socket options in Linux PTP from ETH_P_ALL to ETH_P_1588
observed that the PTP packets are received by Linux PTP only for SAN and HSR.
In case of PRP, the flow got disturbed, and we have made changes in the HSR driver
to allow the PTP packet to be sent to linuxptp in case of PRP.

The fix will return RX_HANDLER_PASS form "hsr_handle_frame" function to
"__netif_receive_skb_core" in case of PRP PTP frames. This ensures that
packets are sent via specific sockets and received by the user application, and
PTP will be in sync.

[Link to PEST board](https://dev.azure.com/PG-BLP/SW/_sprints/taskboard/PEST/SW/PLI22/Sprint%2022.4?workitem=412825)

Related work items: #77574