Merged PR 17895: Wave 3 release 3.9.0
Wave 3 release 3.9.0:
    The Following Enhancement/fix has been included in this release

Linux:
    Fix for the CPU load increase when Linux PTP is running without affecting the tap points

In case of RSTP, the flow got disturbed, and we have made changes in the bridge driver to allow the PTP packet sent to LinuxPTP in case of RSTP.

The fix will return RX_HANDLER_PASS form "br_handle_frame" function to "__netif_receive_skb_core" in case of RSTP PTP frames.
This ensures that packets are sent via specific sockets and received by the user application and PTP will be in sync.

Related work items: #80274