Merged PR 19006: Wave 3 release 3.12.0
Wave 3 release 3.12.0
The following bug fix/feature has been included in this release:

Linux:
REF650: SV and GOOSE messages are not functioning correctly with VLAN ID within the RSTP network

Issue:
The existing VLAN filter in firmware is dumb, and non-PTP Ethernet packets, which
are having a VLAN tag with a different VLAN ID are getting dropped at the firmware level itself.

It is expecting to create VLAN ID entries for all the various VLAN IDs in the network otherwise,
the packets will get dropped by the firmware. This will be tricky, and the filter is not
intelligent enough to learn dynamically.

Fix:
To address this issue, we are disabling the firmware based VLAN filter completely.
Let the bridge or network stack filter the VLAN packets.
From now on, the L2 VLAN filter will always be OFF, in order to let packets receive at the host and then get dropped if they are not intended for the interface.

Changes include disabling the filter and updating features to indicate that we have to filter
the VLAN in software itself rather than dropping them at hardware via PRU.

This may slightly increase CPU load when we are using VLANs only when it is not intended for the DUT.

Related work items: #83154