Merged PR 20087: Wave 3 release 3.15.0
Wave 3 release 3.15.0

The following bug fix/feature has been included in this release:

Driver:
Bug: SV and PTP is not stable in HSR network with maximum devices

Issue:
Observing that Mac rx error are incrementing as we are hitting Rx FIFO size zero with Rx-EOF
Found that Rx-EOF is sticky and not able to clear from the firmware

Fix:
Reset the PHY when the Rx FIFO zero condition is detected.

Implementation:
Implemented a mechanism to reset the PHY when an Rx FIFO lock-up condition is detected.
The Rx FIFO lock-up condition is checked within a hrtimer function, which runs in an atomic context.
When a lock-up is detected, a callback function is invoked in the process context to perform the PHY reset.
A lock mechanism is implemented to ensure that the PHY reset is not triggered again until the ongoing reset is completed.

Please note that this driver change has dependency on the firmware changes
which are posted with the tag "Wave 3 release 3.15.0" on the branch kirkstone of meta-pg-hsa repo.

Related work items: #85406