Merged PR 20637: Wave 3 release 3.16.0
Wave 3 release 3.16.0

Following are the required fixes/workarounds which are included in this release:

Linux:
SV and PTP is not stable in HSR network with maximum devices

Issue:
During reboot tests in the HSR max devices ring, we are creating multiple faults in the HSR ring upon reboots, where continuous error statistics are incremented in few ports, resulting in dropping the packets.

This issue was reproduced only with specific optical SFPs and can be applicable for all protocols since it is a startup issue with SFPs.
During debug, we have found that there is a random optical signal seen on the wire, even if the interface is down, and on the RX side, MII is interpreting this garbage light into symbols, and MII is unable to process this signal, resulting in a large number of error statistics getting incremented.

Debugging at HW and SW levels, we found that the RX Error and False carrier error counters are also incremented, and these are propagated to firmware, and firmware is incrementing MAC RX error counters along with a few different counters like CRC errors and  Misalignment errors.

To fix the issue, we had to make multiple changes across different areas, which are as follows:

Fixes/Workarounds:
1. Change the MDIO interrupt type to level low from edge trigger.
2. Hitachi devices have a way to control SFP via GPIO to make sure we are not observing any garbage data from SFPs during startup
3. The RECR interrupt is disabled to prevent the kernel from getting stuck with servicing these interrupts without clearing them. Added code to read these registers periodically.
4. Control the SFP using GPIO from PRU eth for every interface up and down.

The sequence for controlling SFP is as follows:
a. Disable the SFP before the PHY is stopped during interface down.
b. During interface up, bring the PHY up and then enable the SFP.

More details on the changes made:
DTS Changes
a. Changed the link interrupt type to LEVEL_LOW from EDGE_TRIGGERED
b. Added the SFP disable GPIO pin details to the prueth node so they can be accessed from the PRUETH driver.
c. Initialized the SFP in the OFF state during boot.

PRUETH Driver Changes
a. Check for RECR and FCSCR counters in a periodic timer thread. If these stats are non-zero, then clear bits in the MISR1 Register (this will clear error interrupt bits)
b. Get the SFP GPIO numbers for the corresponding PRU ports from dts.
c. Enable or disable the SFP during interface up and down operations accordingly.

DP83822 PHY driver changes
a. Disabled Receive Error HF interrupt Enable from MISR1 register, so that no error interrupts will be generated for RX ERR events.

Note: The driver changes have a dependency on the firmware changes
which are posted with the tag "Wave 3 release 3.16.0" on the branch "kirkstone" of "meta-pg-hsa" repo.

Related work items: #86998