Merged PR 20816: Wave 3 release 3.17.0
Wave 3 release 3.17.0

Below is the required bug fix which is included in this release:

Linux:
https://dev.azure.com/PG-BLP/BIZ/_workitems/edit/452863 : SVC25-7542 PTP IED Synch Error for merging units in 670 650 IEDs in HSR Network

Issue:
The pdelay RX timestamp is corrupted and resulting in incorrect pdelay computation.

Fix:
Length calculation mismatch was present. This is creating a problem especially during the wraparound of the queue resulting in reading the wrong data as a timestamp.

Related work items: #87524